+++ app/javascript/packs/application.js
@@ -28,25 +28,9 @@ import "bootstrap";
 // Internal imports, e.g:
 // import { initSelect2 } from '../components/init_select2';
 
-document.addEventListener('turbolinks:load', () => {
-  // Call your functions here, e.g:
-  // initSelect2();
-});
-
 // Mapbox
-import 'mapbox-gl/dist/mapbox-gl.css';
-import mapboxgl from 'mapbox-gl';
-mapboxgl.accessToken = 'pk.eyJ1IjoianVua2llc2FuIiwiYSI6ImNrOW83OWg5dTA3YXQzZnBtbWZyYnY5N2EifQ.pSV4ixeAeUHpg6SjgkXoHA';
-const map = new mapboxgl.Map({
-  container: 'map',
-  style: 'mapbox://styles/mapbox/streets-v10'
-});
+import { initMapbox } from '../plugins/init_mapbox';
 
-const mapElement = document.getElementById('map')
-const markers = JSON.parse(mapElement.dataset.markers);
-  markers.forEach((marker) => {
-    new mapboxgl.Marker()
-      .setLngLat([ marker.lng, marker.lat ])
-      .addTo(map);
-  });
-  map.fitBounds(bounds, { padding: 80, maxZoom: 15, duration: 0 });
+document.addEventListener('turbolinks:load', () => {
+  initMapbox();
+})
