+++ app/javascript/plugins/init_mapbox.js
@@ -1,0 +1,34 @@
+import mapboxgl from 'mapbox-gl';
+import MapboxGeocoder from '@mapbox/mapbox-gl-geocoder';
+
+const fitMapToMarkers = (map, markers) => {
+  const bounds = new mapboxgl.LngLatBounds();
+  markers.forEach(marker => bounds.extend([ marker.lng, marker.lat ]));
+  map.fitBounds(bounds, { padding: 80, maxZoom: 15, duration: 0 });
+};
+const resize = (map) => {
+  map.resize();
+};
+const initMapbox = () => {
+  const mapElement = document.getElementById('map');
+  if (mapElement) { // only build a map if there's a div#map to inject into
+    console.log(process.env.MAPBOX_API_KEY);
+    mapboxgl.accessToken = mapElement.dataset.mapboxApiKey;
+    const map = new mapboxgl.Map({
+      container: 'map',
+      style: 'mapbox://styles/mapbox/streets-v10'
+    });
+    window.addEventListener('load', () => { resize(map) })
+    window.addEventListener('resize', () => { resize(map) })
+    const markers = JSON.parse(mapElement.dataset.markers);
+    markers.forEach((marker) => {
+    const popup = new mapboxgl.Popup().setHTML(marker.infoWindow);
+      new mapboxgl.Marker()
+        .setLngLat([ marker.lng, marker.lat ])
+        .setPopup(popup)
+        .addTo(map);
+    });
+    fitMapToMarkers(map, markers);
+  }
+};
+export { initMapbox };
