+++ app/views/trucks/show.html.erb
@@ -87,15 +87,10 @@
       </div>
     </div>
   </div>
-
 <div id="map"
      style="width: 100%; height: 600px;"
      data-markers="<%= @markers.to_json %>"
-     data-mapbox-api-key="<%= ENV['MAPBOX_API_KEY'] %>">
-</div>
-
-
-
+     data-mapbox-api-key="<%= ENV['MAPBOX_API_KEY'] %>"></div>
   <!-- insert Map -->
